Move UnspentOutputsCompararer class to separate file
