+++ Stratis.Bitcoin/Features/Consensus/UInt256Comparer.cs
@@ -14,21 +14,4 @@ namespace Stratis.Bitcoin.Features.Consensus
 			return 0;
 		}
 	}
-	public class UnspentOutputsComparer : IComparer<UnspentOutputs>
-	{
-
-		private static readonly UnspentOutputsComparer _Instance = new UnspentOutputsComparer();
-		public static UnspentOutputsComparer Instance
-		{
-			get
-			{
-				return _Instance;
-			}
-		}
-		private readonly UInt256Comparer Comparer = new UInt256Comparer();
-		public int Compare(UnspentOutputs x, UnspentOutputs y)
-		{
-			return this.Comparer.Compare(x.TransactionId, y.TransactionId);
-		}
-	}
 }
